requirements.txt
RPi.GPIO>=0.7.0
Flask>=2.0.0

